requirements.txt
aiohttp
requests
psycopg2-binary
discord-py-interactions==5.6.0